# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0

paths:
  /tasks:
    get:
      summary: Get all tasks
      tags:
        - tasks
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      tags:
        - tasks
      requestBody:
        description: The task to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: The created task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

  /users/{id}/tasks:
    get:
      summary: Get tasks by user ID
      tags:
        - tasks
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID пользователя
      responses:
        '200':
          description: Список задач пользователя
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '400':
          description: Некорректный ID
        '500':
          description: Ошибка сервера

  /tasks/{id}:
    patch:
      summary: Обновить задачу по ID
      tags:
        - tasks
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Уникальный идентификатор задачи
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: Обновлённая задача
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Некорректный ID или данные
        '500':
          description: Ошибка сервера

    delete:
      summary: Удалить задачу по ID
      tags:
        - tasks
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Уникальный идентификатор задачи
      responses:
        '204':
          description: Задача успешно удалена
        '400':
          description: Некорректный ID
        '500':
          description: Ошибка сервера

  /users:
    get:
      summary: Get all users
      tags:
        - users
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'

    post:
      summary: Create a new user
      tags:
        - users
      requestBody:
        description: The user to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: The created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /users/{id}:
    patch:
      summary: Обновить пользователя по ID
      tags:
        - users
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Уникальный идентификатор пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Обновлённый пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Некорректный ID или данные
        '500':
          description: Ошибка сервера

    delete:
      summary: Удалить пользователя по ID
      tags:
        - users
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Уникальный идентификатор пользователя
      responses:
        '204':
          description: Пользователь успешно удалён
        '400':
          description: Некорректный ID
        '500':
          description: Ошибка сервера

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          format: uint
        task:
          type: string
        is_done:
          type: boolean
        user_id:
          type: integer
          format: uint
    User:
      type: object
      properties:
        id:
          type: integer
          format: uint
        email:
          type: string
        password:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true